experimental/pokerbots/requirements.txt
functions-framework==3.2.1
flask==2.1.0
google-cloud-storage==2.7.0
